@@ -1,26 +1,54 @@
-exports.L = require('leaflet');
-exports.findRealParent = require('./utils/findRealParent').default;
-exports.propsBinder = require('./utils/propsBinder').default;
+import L from 'leaflet';
+import findRealParent from './utils/findRealParent';
+import propsBinder from './utils/propsBinder';
 
-exports.LCircle = require('./components/LCircle').default;
-exports.LCircleMarker = require('./components/LCircleMarker').default;
-exports.LControl = require('./components/LControl').default;
-exports.LControlAttribution = require('./components/LControlAttribution').default;
-exports.LControlLayers = require('./components/LControlLayers').default;
-exports.LControlScale = require('./components/LControlScale').default;
-exports.LControlZoom = require('./components/LControlZoom').default;
-exports.LFeatureGroup = require('./components/LFeatureGroup').default;
-exports.LGeoJson = require('./components/LGeoJson').default;
-exports.LIcon = require('./components/LIcon').default;
-exports.LIconDefault = require('./components/LIconDefault').default;
-exports.LImageOverlay = require('./components/LImageOverlay').default;
-exports.LLayerGroup = require('./components/LLayerGroup').default;
-exports.LMap = require('./components/LMap').default;
-exports.LMarker = require('./components/LMarker').default;
-exports.LPolygon = require('./components/LPolygon').default;
-exports.LPolyline = require('./components/LPolyline').default;
-exports.LPopup = require('./components/LPopup').default;
-exports.LRectangle = require('./components/LRectangle').default;
-exports.LTileLayer = require('./components/LTileLayer').default;
-exports.LTooltip = require('./components/LTooltip').default;
-exports.LWMSTileLayer = require('./components/LWMSTileLayer').default;
+import LCircle from './components/LCircle';
+import LCircleMarker from './components/LCircleMarker';
+import LControl from './components/LControl';
+import LControlAttribution from './components/LControlAttribution';
+import LControlLayers from './components/LControlLayers';
+import LControlScale from './components/LControlScale';
+import LControlZoom from './components/LControlZoom';
+import LFeatureGroup from './components/LFeatureGroup';
+import LGeoJson from './components/LGeoJson';
+import LIcon from './components/LIcon';
+import LIconDefault from './components/LIconDefault';
+import LImageOverlay from './components/LImageOverlay';
+import LLayerGroup from './components/LLayerGroup';
+import LMap from './components/LMap';
+import LMarker from './components/LMarker';
+import LPolygon from './components/LPolygon';
+import LPolyline from './components/LPolyline';
+import LPopup from './components/LPopup';
+import LRectangle from './components/LRectangle';
+import LTileLayer from './components/LTileLayer';
+import LTooltip from './components/LTooltip';
+import LWMSTileLayer from './components/LWMSTileLayer';
+
+export {
+  L,
+  findRealParent,
+  propsBinder,
+  LCircle,
+  LCircleMarker,
+  LControl,
+  LControlAttribution,
+  LControlLayers,
+  LControlScale,
+  LControlZoom,
+  LFeatureGroup,
+  LGeoJson,
+  LIcon,
+  LIconDefault,
+  LImageOverlay,
+  LLayerGroup,
+  LMap,
+  LMarker,
+  LPolygon,
+  LPolyline,
+  LPopup,
+  LRectangle,
+  LTileLayer,
+  LTooltip,
+  LWMSTileLayer
+};
