@@ -1,6 +1,6 @@
 <template>
   <div>
-    <slot/>
+    <slot />
   </div>
 </template>
 
