@@ -83,7 +83,7 @@ export default {
   },
   render: function (h) {
     if (this.ready && this.$slots.default) {
-      return h('div', {style: {display: 'none'}}, this.$slots.default);
+      return h('div', { style: { display: 'none' } }, this.$slots.default);
     }
     return null;
   }
